webapp/requirements.txt
flask==2.2.3
werkzeug==2.2.3
prometheus-client==0.11.0
twilio==7.17.0
python-dotenv==1.0.0 